+++ src/mgard-x/CMakeLists.txt
@@ -9,6 +9,7 @@ add_subdirectory (CompressionHighLevel)
 if (MGARD_ENABLE_MDR)
   add_subdirectory (MDRHighLevel)
 endif()
+add_subdirectory (Metadata)
 add_subdirectory (RuntimeX)
 set(MGARD_X_SERIAL_SRC ${MGARD_X_SERIAL_SRC} PARENT_SCOPE)
 set(MGARD_X_OPENMP_SRC ${MGARD_X_OPENMP_SRC} PARENT_SCOPE)
