+++ src/mgard-x/Metadata/CMakeLists.txt
@@ -1,0 +1,4 @@
+list(APPEND MGARD_X_SRC  
+      ${CMAKE_CURRENT_SOURCE_DIR}/Metadata.cpp)
+
+set(MGARD_X_SRC ${MGARD_X_SRC} PARENT_SCOPE)
